+++ packages/sync-server/jest.setup.ts
@@ -1,0 +1,7 @@
+import { getTestDatabase, clearTestData } from '@tupaia/database';
+
+afterAll(async () => {
+  const database = getTestDatabase();
+  await clearTestData(database);
+  await database.closeConnections();
+});
